separated the menu component
